id_tool-camera.cpp - identify a tool in a viewport
all this code does is look through the camera and a small ROI in an
attempt to identify a tool. it is up to the caller to put the arm in
the correct position.

as it stands, id_flame-camera and id_tool-camera both leave windows
open so that the user can see the end result.