modified commander and subjects to fix ros init
ROS requires that ros::init() be called *before* a nodehandle is
created. altered the class constructors to use pointers to a node-
handle that is created in commander/main.cpp and then passed amongst
the objects as they are created. this works.

it appears that when calling execute binary that it launches the file,
but in a separate environment and so no I/O through ROS_INFO is reaching
the launching console.