implementing new move function
instead of using the inverse kinematic to put the arm to a point, we
take the requested position passwd across /arm/put/point and send that
to a new fucntion: put_point_line(msg);

this function finds the largest difference between the two points and
uses that to guide the arm along that long line. handles increasing/
decreasing, and guide being on any of the three axes.