debuggin move_line
arm is twitchy - i think this has to do with the initial_park and the
co-ordinates i'm telling it that it is at. any point i give it after-
wards wigs out due to the origin being so close, the arm can't handle
the quick change in location correctly.

trying to add back in the query function and the rosserial_python bridge
is yelling about short reads.

rolling back to a previous, known working, copy of the code to verify
that the board isn't fried.